# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marketplace de Servicios API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: List all active listings
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [service, product, null]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: locality
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingWithProvider"

    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing (providers only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingWithProvider"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingBody"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"

    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /upload/image:
    post:
      operationId: uploadImage
      tags: [listings]
      summary: Upload an image and get URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadImageBody"
      responses:
        "200":
          description: Image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadImageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [provider, client, admin, company]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        notifyEmail:
          type: boolean
        companyName:
          type: ["string", "null"]
        cuit:
          type: ["string", "null"]
        companyAddress:
          type: ["string", "null"]
        companyIndustry:
          type: ["string", "null"]
        companyApproved:
          type: boolean
        cvUrl:
          type: ["string", "null"]
        cvPublic:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - notifyEmail
        - companyApproved
        - cvPublic
        - createdAt
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: ["string", "null"]
        type:
          type: string
          enum: [service, product]
        description:
          type: ["string", "null"]
      required:
        - id
        - name
        - type
    Listing:
      type: object
      properties:
        id:
          type: integer
        providerId:
          type: integer
        categoryId:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [service, product]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        whatsapp:
          type: ["string", "null"]
        paymentMethods:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        adminApproved:
          type: boolean
        quantity:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, sold, paused]
        pricingType:
          type: string
          enum: [unit, per_kilo]
        weightKg:
          type: ["number", "null"]
        sizes:
          type: ["array", "null"]
          items:
            type: object
            properties:
              name:
                type: string
              price:
                type: number
              stock:
                type: ["integer", "null"]
            required:
              - name
              - price
        variantLabel:
          type: ["string", "null"]
        requiresSchedule:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - providerId
        - categoryId
        - title
        - description
        - type
        - price
        - paymentMethods
        - isActive
        - adminApproved
        - status
        - pricingType
        - requiresSchedule
        - createdAt
...
    ListingWithProvider:
      type: object
      properties:
        id:
          type: integer
        providerId:
          type: integer
        categoryId:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [service, product]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        whatsapp:
          type: ["string", "null"]
        paymentMethods:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        adminApproved:
          type: boolean
        quantity:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, sold, paused]
        pricingType:
          type: string
          enum: [unit, per_kilo]
        weightKg:
          type: ["number", "null"]
        sizes:
          type: ["array", "null"]
          items:
            type: object
            properties:
              name:
                type: string
              price:
                type: number
              stock:
                type: ["integer", "null"]
            required:
              - name
              - price
        variantLabel:
          type: ["string", "null"]
        requiresSchedule:
          type: boolean
        createdAt:
          type: string
        provider:
          $ref: "#/components/schemas/User"
        category:
          $ref: "#/components/schemas/Category"
      required:
        - id
        - providerId
        - categoryId
        - title
        - description
        - type
        - price
        - paymentMethods
        - isActive
        - adminApproved
        - status
        - pricingType
        - requiresSchedule
        - createdAt
        - provider
        - category
    CreateListingBody:
      type: object
      properties:
        categoryId:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [service, product]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        whatsapp:
          type: ["string", "null"]
        paymentMethods:
          type: array
          items:
            type: string
        quantity:
          type: ["integer", "null"]
        pricingType:
          type: string
          enum: [unit, per_kilo]
        weightKg:
          type: ["number", "null"]
        sizes:
          type: ["array", "null"]
          items:
            type: object
            properties:
              name:
                type: string
              price:
                type: number
              stock:
                type: ["integer", "null"]
            required:
              - name
              - price
        variantLabel:
          type: ["string", "null"]
        requiresSchedule:
          type: ["boolean", "null"]
      required:
        - categoryId
        - title
        - description
        - type
        - price
        - paymentMethods
    UpdateListingBody:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
          enum: [service, product, null]
        price:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: ["array", "null"]
          items:
            type: string
        whatsapp:
          type: ["string", "null"]
        paymentMethods:
          type: ["array", "null"]
          items:
            type: string
        isActive:
          type: ["boolean", "null"]
        quantity:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
          enum: [active, sold, paused, null]
        pricingType:
          type: ["string", "null"]
          enum: [unit, per_kilo, null]
        weightKg:
          type: ["number", "null"]
        sizes:
          type: ["array", "null"]
          items:
            type: object
            properties:
              name:
                type: string
              price:
                type: number
              stock:
                type: ["integer", "null"]
            required:
              - name
              - price
        variantLabel:
          type: ["string", "null"]
        requiresSchedule:
          type: ["boolean", "null"]
...
    UploadImageBody:
      type: object
      properties:
        base64:
          type: string
        filename:
          type: string
      required:
        - base64
        - filename
    UploadImageResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...