# lib/api-zod/src/generated/types/listingWithProvider.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketplace de Servicios API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";
import type { ListingWithProviderPricingType } from "./listingWithProviderPricingType";
import type { ListingWithProviderSizesItem } from "./listingWithProviderSizesItem";
import type { ListingWithProviderStatus } from "./listingWithProviderStatus";
import type { ListingWithProviderType } from "./listingWithProviderType";
import type { User } from "./user";

export interface ListingWithProvider {
  id: number;
  providerId: number;
  categoryId: number;
  title: string;
  description: string;
  type: ListingWithProviderType;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  /** @nullable */
  whatsapp?: string | null;
  paymentMethods: string[];
  isActive: boolean;
  adminApproved: boolean;
  /** @nullable */
  quantity?: number | null;
  status: ListingWithProviderStatus;
  pricingType: ListingWithProviderPricingType;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  sizes?: ListingWithProviderSizesItem[] | null;
  /** @nullable */
  variantLabel?: string | null;
  requiresSchedule: boolean;
  createdAt: string;
  provider: User;
  category: Category;
}